requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.5
Deprecated==1.2.14
dnspython==2.2.1
email-validator==1.3.0
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cache==0.13.1
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
geographiclib==2.0
geopy==2.3.0
greenlet==1.1.3.post0
idna==3.4
imgurpython==1.1.7
importlib-resources==5.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
limits==3.5.0
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
ordered-set==4.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
Pygments==2.15.1
PyJWT==2.5.0
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.18.1
pytest==7.1.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.4
qrcode==7.4.2
requests==2.31.0
rich==13.4.2
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
tomli==2.0.1
typing_extensions==4.6.3
ua-parser==0.16.1
urllib3==2.0.3
user-agents==2.2.0
Werkzeug==2.1.2
wrapt==1.15.0
